refactor review v1 page
